Create Camera.py
This object takes in the dev/video# of a camera linked to the system via a usb. The current version goes through 3 steps, (1) the user defines the dev/video# and the stream width. The class then gets the intrinsic height of the camera. (2) The user then confirms if the camera is calibrated for the displace aspect ratio. (3) The user then is prompted to calibrate the camera or to recalibrate the camera. This is done specifically because the tracking requires the calibration mtx and dist.