Update Tracker.py
A thorough update to the tracking code which focused on uncomplicating the arucoDetect function to allow for multifunction usage of it. 
The code was also geared specifically for the project for which general usage was ignored. The code also increased in size to allow for more in-depth projectTracker and getCameraRelations2Global algorithms.